third_party/WebKit/LayoutTests/inspector/tracing/timeline-timer.html

<html>
<head>
<script src="../../http/tests/inspector/inspector-test.js"></script>
<script src="../../http/tests/inspector/timeline-test.js"></script>
<script>

function performActions(callback)
{
    var timerOne = setTimeout("1 + 1", 10);
    var timerTwo = setInterval(intervalTimerWork, 20);
    var iteration = 0;

    function intervalTimerWork()
    {
        if (++iteration < 2)
            return;
        clearInterval(timerTwo);
        callback();
    }
}

function test()
{
    WebInspector.panels.timeline._captureJSProfileSetting.set(false);
    InspectorTest.invokeAsyncWithTimeline("performActions", finish);

    function finish()
    {
        InspectorTest.printTimelineRecordsWithDetails("TimerInstall");
        InspectorTest.printTimelineRecordsWithDetails("TimerFire");
        InspectorTest.printTimelineRecordsWithDetails("TimerRemove");
        InspectorTest.printTimelineRecords("FunctionCall");
        InspectorTest.printTimelineRecordsWithDetails("EvaluateScript");
        InspectorTest.completeTest();
    }
}

if (!window.testRunner)
    setTimeout(performActions, 2000);

</script>
</head>

<body onload="runTest()">
<p>
Tests the Timeline events for Timers
</p>
</body>
</html>
